refactor: (AHWR-2144) commit pact contract instead of publishing via GitHub Release #minor
Also strengthens the contract test itself (exact total assertion, explicit orphaned-claim
absence check) and renames the claims fixture to claims-response.js for clarity.
