+++ .github/workflows/publish.yml
@@ -32,25 +32,12 @@ jobs:
           npm install -g npm@11.10.0
           ./scripts/test
 
-      - name: Publish Pact contract
-        id: publish-pact
-        if: success()
-        continue-on-error: true
+      - name: Verify committed pact matches generated contract
         run: |
-          gh release download pact-contracts --pattern "*.json" --dir /tmp/existing-pact --clobber 2>/dev/null || true
-          existing_file=$(ls /tmp/existing-pact/*.json 2>/dev/null | head -1)
-
-          if [ -n "$existing_file" ] && diff -q <(jq -S . "$existing_file") <(jq -S . pacts/*.json) > /dev/null 2>&1; then
-            echo "Contract unchanged since last publish - nothing to do"
-          else
-            echo "Contract changed (or no previous release exists) - publishing"
-            gh release delete pact-contracts --yes || true
-            gh release create pact-contracts pacts/*.json \
-              --title "Pact contracts (latest)" \
-              --notes "Auto-published from ${{ github.sha }}"
+          if [ -n "$(git status --porcelain pacts/)" ]; then
+            echo "::error::Committed pact is stale - rerun the contract test and commit the regenerated pacts/*.json"
+            exit 1
           fi
-        env:
-          GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
 
       - name: Build and Publish
         uses: DEFRA/cdp-build-action/build@main
@@ -71,9 +58,3 @@ jobs:
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           SONAR_TOKEN: ${{ secrets.SONAR_TOKEN }}
-
-      - name: Fail if Pact contract publish failed
-        if: steps.publish-pact.outcome == 'failure'
-        run: |
-          echo "::error::Pact contract publish failed - ahwr-application-backend may verify against a stale contract until this is fixed."
-          exit 1
