+++ .gitignore
@@ -1,7 +1,6 @@
 # Test output
 html-reports
 test-output
-pacts
 
 # npm stuff
 .npm/
