+++ .prettierignore
@@ -1,5 +1,6 @@
 test-output
 app/frontend/dist
+pacts
 snyk-cli
 snyk_report.html
 snyk-result.json
